refactor: use new consent computation
