--- src/components/EstimationSummary.test.js
@@ -1,53 +1,0 @@
-import { calculateConsensus } from "./EstimationSummary";
-
-// Generated by CodiumAI
-
-describe("calculateConsensus", () => {
-  it("calculates consensus for an evenly distributed range", () => {
-    const estimations = ["1", "2", "3"];
-    const result = calculateConsensus(estimations);
-    expect(result).toBe(2);
-  });
-
-  it("handles consensus when all estimations are identical", () => {
-    const estimations = ["2", "2", "2"];
-    const result = calculateConsensus(estimations);
-    expect(result).toBe(2);
-  });
-
-  it("computes consensus with a single estimation", () => {
-    const estimations = ["2"];
-    const result = calculateConsensus(estimations);
-    expect(result).toBe(2);
-  });
-
-  it("determines consensus for a pair with an average needing rounding up", () => {
-    const estimations = ["1", "2"];
-    const result = calculateConsensus(estimations);
-    expect(result).toBe(2);
-  });
-
-  it("finds consensus in a pair with an exact average match", () => {
-    const estimations = ["1", "3"];
-    const result = calculateConsensus(estimations);
-    expect(result).toBe(2);
-  });
-
-  it("selects the nearest higher value for consensus between two widely spaced estimations", () => {
-    const estimations = ["3", "8"];
-    const result = calculateConsensus(estimations);
-    expect(result).toBe(5);
-  });
-
-  it("chooses the nearest higher consensus in an imbalanced set of estimations", () => {
-    const estimations = ["21", "1", "1"];
-    const result = calculateConsensus(estimations);
-    expect(result).toBe(8);
-  });
-
-  it("selects the closest consensus value in a balanced set with repeated values", () => {
-    const estimations = ["5", "5", "5", "13", "13"];
-    const result = calculateConsensus(estimations);
-    expect(result).toBe(8);
-  });
-});
